+++ src/world/objects/Disk.cpp
@@ -14,4 +14,8 @@ std::shared_ptr<render::Primitive> Disk::toRaytracerPrimitive() const {
   return make_named<render::Disk>(Vector3d::null, Vector3d(0, 1, 0), m_radius);
 }
 
+bool Disk::supportsPlanarSceneMarker() const {
+  return true;
+}
+
 static bool dummy = ElementFactory::self().registerClass<Disk>("Disk");